Report capacity progression through a Ready condition
We're moving towards relocating the smarts in the checks for a target's
readiness away from the executor and into the controllers themselves,
leaving the executor to check for a single `Ready` condition at the
target's top level status. This commit does exactly that for
CapacityTarget objects.

+++ pkg/controller/release/strategy_executor.go
@@ -128,7 +128,7 @@ func genCapacityEnforcer(curr, succ *releaseInfo) PipelineStep {
 
 		isLastStep := int(targetStep) == len(strategy.Steps)-1
 
-		if achieved, newSpec, clustersNotReady := checkCapacity(curr.capacityTarget, uint(capacityWeight)); !achieved {
+		if achieved, newSpec, clustersNotReady := checkCapacity(curr.capacityTarget, capacityWeight); !achieved {
 			e.info("release hasn't achieved capacity yet")
 
 			var patches []ExecutorResult
